refactor: 카테고리 custom converter 추가
