+++ backend/http/post.http
@@ -4,10 +4,21 @@ Authorization: Bearer {{ access_token }}
 Content-Type: application/json
 
 {
-  "title": {{ title }},
-  "content" : {{ content }},
-  "price" : {{ price }},
-  "category" : {{ category }},
-  "thumbnail" : {{ thumbnail }},
-  "imageRequest" : {{imageRequest}}
+  "title": "제목입니다.",
+  "content": "내용입니다.",
+  "price": 100000,
+  "category": "디지털기기",
+  "thumbnail": "tumbnail.com",
+  "imageRequest": [
+    {
+      "imageUrl": "http://tumbnail.com/1",
+      "originName": "사진1"
+    },
+    {
+      "imageUrl": "http://tumbnail.com/2",
+      "originName": "사진2"
+    }
+  ]
 }
+
+
