+++ backend/src/main/java/com/carrot/application/post/service/PostWriteService.java
@@ -16,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.carrot.presentation.request.PostDto.*;
+import static com.carrot.presentation.request.PostRequest.*;
 
 @Service
 @Transactional
